Major changes - Change UI design - Install flask
