+++ .gitignore
@@ -5,3 +5,5 @@
 **/*.log
 #Temp File
 tempfiles/
+
+env/
