+++ old_web/js/global.js
