+++ requirements.txt
@@ -1,0 +1,179 @@
+alabaster==0.7.8
+alembic==0.9.3.dev0
+apturl==0.5.2
+asgiref==3.2.10
+asn1crypto==0.24.0
+astroid==2.3.3
+autopep8==1.5
+Babel==2.4.0
+backcall==0.1.0
+beautifulsoup4==4.6.0
+blinker==1.4
+bottle==0.12.17
+bottle-websocket==0.2.9
+Brlapi==0.6.6
+certifi==2019.6.16
+cffi==1.12.3
+chardet==3.0.4
+click==6.7
+colorama==0.3.7
+command-not-found==0.3
+cryptography==2.7
+cupshelpers==1.0
+cycler==0.10.0
+decorator==4.4.1
+defer==1.0.6
+distro-info===0.18ubuntu0.18.04.1
+Django==2.2.4
+django-autocomplete-light==3.3.5
+django-bootstrap4==0.0.8
+django-cors-headers==3.7.0
+django-crispy-forms==1.7.2
+django-filter==2.1.0
+django-guardian==2.0.0
+django-mathfilters==0.4.0
+django-nose==1.4.6
+django-queryset-csv==1.0.2
+django-rest-auth==0.9.5
+django-rest-knox==4.1.0
+django-tables2==2.1.1
+djangorestframework==3.9.4
+docutils==0.14
+Eel==0.11.0
+Flask==1.0
+Flask-BabelEx==0.9.3
+Flask-Compress==1.4.0
+Flask-Gravatar==0.4.2
+Flask-Login==0.4.0
+Flask-Mail==0.9.1
+Flask-Migrate==2.1.1
+Flask-Paranoid==0.2.0
+Flask-Principal==0.4.0
+Flask-Security-Too==3.4.2
+Flask-SQLAlchemy==2.1
+Flask-WTF==0.14.2
+future==0.18.2
+GDAL==2.2.2
+gevent==1.4.0
+gevent-websocket==0.10.1
+greenlet==0.4.15
+gunicorn==20.1.0
+html5lib==0.999999999
+httplib2==0.11.3
+idna==2.8
+imagesize==0.7.1
+ipython==7.9.0
+ipython-genutils==0.2.0
+iso8601==0.1.12
+isort==4.3.21
+itsdangerous==0.24
+jedi==0.15.1
+Jinja2==2.11.3
+jsonschema==2.6.0
+jupyter-core==4.4.0
+kazam==1.4.5
+keyring==10.6.0
+keyrings.alt==3.0
+language-selector==0.1
+launchpadlib==1.10.6
+lazr.restfulclient==0.13.5
+lazr.uri==1.0.3
+lazy-object-proxy==1.4.3
+ldap3==2.4.1
+louis==3.5.0
+lxml==4.2.1
+macaroonbakery==1.1.3
+Mako==1.0.7
+Markdown==3.1.1
+MarkupSafe==2.0.1
+matplotlib==2.1.1
+mccabe==0.6.1
+nbformat==4.4.0
+netifaces==0.10.4
+nose==1.3.7
+numpy==1.13.3
+oauth==1.0.1
+olefile==0.45.1
+OWSLib==0.16.0
+paramiko==2.0.0
+parso==0.5.1
+passlib==1.7.1
+pathtools==0.1.2
+pexpect==4.2.1
+pickleshare==0.7.5
+Pillow==5.1.0
+plotly==2.2.3
+prompt-toolkit==2.0.10
+protobuf==3.0.0
+psutil==5.4.2
+psycopg2==2.8.6
+pyasn1==0.4.2
+pycairo==1.16.2
+pycodestyle==2.5.0
+pycparser==2.19
+pycrypto==2.6.1
+pycups==1.9.73
+PyFladesk==1.1
+Pygments==2.4.2
+PyGObject==3.26.1
+pyinotify==0.9.6
+pylint==2.4.4
+pymacaroons==0.13.0
+PyNaCl==1.1.2
+pyOpenSSL==17.5.0
+pyparsing==2.2.0
+pyproj==1.9.5.1
+PyQt5==5.15.0
+PyQt5-Qt5==5.15.2
+PyQt5-sip==12.8.1
+PyQtWebEngine==5.15.2
+PyQtWebEngine-Qt5==5.15.2
+pyRFC3339==1.0
+python-apt==1.6.5+ubuntu0.6
+python-dateutil==2.6.1
+python-debian==0.1.32
+pytz==2019.1
+pyxdg==0.25
+PyYAML==3.12
+reportlab==3.4.0
+requests==2.22.0
+requests-unixsocket==0.1.5
+roman==2.0.0
+scapy==2.4.3
+SecretStorage==2.3.1
+simplejson==3.13.2
+six==1.12.0
+slip==0.6.5
+slip.dbus==0.6.5
+snowflake==0.0.3
+speaklater==1.4
+Sphinx==1.6.7
+SQLAlchemy==1.1.11
+sqlparse==0.3.0
+ssh-import-id==5.7
+sshtunnel==0.1.4
+system-service==0.3
+systemd-python==234
+traitlets==4.3.3
+typed-ast==1.4.1
+ubuntu-advantage-tools==27.0
+ubuntu-drivers-common==0.0.0
+ufw==0.36
+unattended-upgrades==0.1
+unicodecsv==0.14.1
+urllib3==1.25.3
+usb-creator==0.3.3
+virtualenv==15.1.0
+wadllib==1.3.2
+watchdog==0.10.0
+wcwidth==0.1.7
+webencodings==0.5
+websockets==8.1
+WebUI @ git+git://github.com/widdershin/flask-desktop.git@8ce810284c2b869594893b059862d3ef66bb5f61
+Werkzeug==0.15.5
+wget==3.2
+whichcraft==0.6.1
+wrapt==1.11.2
+WTForms==2.1
+xkit==0.0.0
+zope.interface==4.3.2
