+++ web/static/css/home.css
@@ -1,0 +1,8 @@
+@font-face {
+  font-family: Roboto, sans-serif;
+  src: ('Roboto-Regular.ttf') format('truetype');
+}
+
+body, html {
+  font-size: Roboto, sans-serif !important;
+}
